### 2021-12-03 #3975 @whedon
[PRE REVIEW]: ADaPT-ML: A Data Programming Template for Machine Learning
**Submitting author:** @nulberry (<a href="http://orcid.org/0000-0001-8512-1810">Andrea Whittaker</a>)
**Repository:** <a href="https://github.com/U-Alberta/ADaPT-ML" target ="_blank">https://github.com/U-Alberta/ADaPT-ML</a>
**Version:** v1.0.0
