### 2021-12-06 #3975 @jmschrei
@aaronpeikert are you able to review this submission? The GitHub is mostly in Python but the tool seems to not involve programming that much.

@julianpistorius are you able to review this submission?
