### 2022-01-07 #4038 @whedon
[REVIEW]: ADaPT-ML: A Data Programming Template for Machine Learning
**Submitting author:** @nulberry (<a href="http://orcid.org/0000-0001-8512-1810">Andrea Whittaker</a>)
**Repository:** https://github.com/U-Alberta/ADaPT-ML
**Branch with paper.md** (empty if default branch): main
**Version:** v1.0.0

- [x] #4076
