### 2022-01-24 U-Alberta/ADaPT-ML#9 @aaronpeikert
docker compose connection issues
I tried to run `./test/all-unix.sh` without much success.

`bash test/all-unix.sh` results in

<pre>=== STARTING ADaPT-ML AND RUNNING ALL TESTS ===
label-studio-web is up-to-date
Starting modelling-mlflow-db ...
Starting modelling-mlflow-db     ... <font color="#26A269">done</font>
crate-db is up-to-date
Starting dp-mlflow-db            ... <font color="#26A269">done</font>
Starting modelling-mlflow-deploy ...
Starting label-studio-dev        ... <font color="#26A269">done</font>
Starting modelling-mlflow-deploy ... <font color="#26A269">done</font>
Starting modelling-mlflow        ... <font color="#26A269">done</font>
Starting dp-mlflow               ... <font color="#26A269">done</font>
dp-mlflow-db              /entrypoint.sh mysqld            Up (health: starting)   3306/tcp, 33060/tcp, 33061/tcp
modelling-mlflow-db       /entrypoint.sh mysqld            Up (health: starting)   3306/tcp, 33060/tcp, 33061/tcp
Waiting for MLflow databases (5 sec)...
dp-mlflow-db              /entrypoint.sh mysqld            Up (health: starting)   3306/tcp, 33060/tcp, 33061/tcp
modelling-mlflow-db       /entrypoint.sh mysqld            Up (health: starting)   3306/tcp, 33060/tcp, 33061/tcp
Waiting for MLflow databases (5 sec)...
dp-mlflow-db              /entrypoint.sh mysqld            Up (health: starting)   3306/tcp, 33060/tcp, 33061/tcp
modelling-mlflow-db       /entrypoint.sh mysqld            Up (health: starting)   3306/tcp, 33060/tcp, 33061/tcp
Waiting for MLflow databases (5 sec)...
dp-mlflow-db              /entrypoint.sh mysqld            Up (health: starting)   3306/tcp, 33060/tcp, 33061/tcp
modelling-mlflow-db       /entrypoint.sh mysqld            Up (health: starting)   3306/tcp, 33060/tcp, 33061/tcp
Waiting for MLflow databases (5 sec)...
crate-db                  /docker-entrypoint.sh crat ...   Restarting
dp-mlflow-db              /entrypoint.sh mysqld            Up (health: starting)   3306/tcp, 33060/tcp, 33061/tcp
modelling-mlflow-db       /entrypoint.sh mysqld            Up (health: starting)   3306/tcp, 33060/tcp, 33061/tcp
Waiting for MLflow databases (5 sec)...
crate-db                  /docker-entrypoint.sh crat ...   Restarting
Waiting for MLflow databases (5 sec)...
Checking for exited containers...
Startup complete.
Traceback (most recent call last):
  File &quot;/usr/local/lib/python3.8/site-packages/urllib3/connection.py&quot;, line 174, in _new_conn
    conn = connection.create_connection(
  File &quot;/usr/local/lib/python3.8/site-packages/urllib3/util/connection.py&quot;, line 73, in create_connection
    for res in socket.getaddrinfo(host, port, family, socket.SOCK_STREAM):
  File &quot;/usr/local/lib/python3.8/socket.py&quot;, line 918, in getaddrinfo
    for res in _socket.getaddrinfo(host, port, family, type, proto, flags):
socket.gaierror: [Errno -2] Name or service not known

Traceback (most recent call last):
  File &quot;/usr/local/lib/python3.8/site-packages/urllib3/connectionpool.py&quot;, line 699, in urlopen
    httplib_response = self._make_request(
  File &quot;/usr/local/lib/python3.8/site-packages/urllib3/connectionpool.py&quot;, line 394, in _make_request
    conn.request(method, url, **httplib_request_kw)
  File &quot;/usr/local/lib/python3.8/site-packages/urllib3/connection.py&quot;, line 239, in request
    super(HTTPConnection, self).request(method, url, body=body, headers=headers)
  File &quot;/usr/local/lib/python3.8/http/client.py&quot;, line 1256, in request
    self._send_request(method, url, body, headers, encode_chunked)
  File &quot;/usr/local/lib/python3.8/http/client.py&quot;, line 1302, in _send_request
    self.endheaders(body, encode_chunked=encode_chunked)
  File &quot;/usr/local/lib/python3.8/http/client.py&quot;, line 1251, in endheaders
    self._send_output(message_body, encode_chunked=encode_chunked)
  File &quot;/usr/local/lib/python3.8/http/client.py&quot;, line 1011, in _send_output
    self.send(msg)
  File &quot;/usr/local/lib/python3.8/http/client.py&quot;, line 951, in send
    self.connect()
  File &quot;/usr/local/lib/python3.8/site-packages/urllib3/connection.py&quot;, line 205, in connect
    conn = self._new_conn()
  File &quot;/usr/local/lib/python3.8/site-packages/urllib3/connection.py&quot;, line 186, in _new_conn
    raise NewConnectionError(
urllib3.exceptions.NewConnectionError: &lt;urllib3.connection.HTTPConnection object at 0x7f29a923d340&gt;: Failed to establish a new connection: [Errno -2] Name or service not known

Traceback (most recent call last):
  File &quot;/usr/local/lib/python3.8/site-packages/crate/client/http.py&quot;, line 473, in _request
    response = self.server_pool[next_server].request(
  File &quot;/usr/local/lib/python3.8/site-packages/crate/client/http.py&quot;, line 154, in request
    return self.pool.urlopen(
  File &quot;/usr/local/lib/python3.8/site-packages/urllib3/connectionpool.py&quot;, line 783, in urlopen
    return self.urlopen(
  File &quot;/usr/local/lib/python3.8/site-packages/urllib3/connectionpool.py&quot;, line 783, in urlopen
    return self.urlopen(
  File &quot;/usr/local/lib/python3.8/site-packages/urllib3/connectionpool.py&quot;, line 783, in urlopen
    return self.urlopen(
  [Previous line repeated 7 more times]
  File &quot;/usr/local/lib/python3.8/site-packages/urllib3/connectionpool.py&quot;, line 755, in urlopen
    retries = retries.increment(
  File &quot;/usr/local/lib/python3.8/site-packages/urllib3/util/retry.py&quot;, line 574, in increment
    raise MaxRetryError(_pool, url, error or ResponseError(cause))
urllib3.exceptions.MaxRetryError: HTTPConnectionPool(host=&apos;crate-db&apos;, port=4200): Max retries exceeded with url: /_sql (Caused by NewConnectionError(&apos;&lt;urllib3.connection.HTTPConnection object at 0x7f29a923d340&gt;: Failed to establish a new connection: [Errno -2] Name or service not known&apos;))

Traceback (most recent call last):
  File &quot;/usr/local/lib/python3.8/site-packages/sqlalchemy/engine/base.py&quot;, line 1802, in _execute_context
    self.dialect.do_execute(
  File &quot;/usr/local/lib/python3.8/site-packages/sqlalchemy/engine/default.py&quot;, line 719, in do_execute
    cursor.execute(statement, parameters)
  File &quot;/usr/local/lib/python3.8/site-packages/crate/client/cursor.py&quot;, line 53, in execute
    self._result = self.connection.client.sql(sql, parameters,
  File &quot;/usr/local/lib/python3.8/site-packages/crate/client/http.py&quot;, line 396, in sql
    content = self._json_request(&apos;POST&apos;, self.path, data=data)
  File &quot;/usr/local/lib/python3.8/site-packages/crate/client/http.py&quot;, line 522, in _json_request
    response = self._request(method, path, data=data)
  File &quot;/usr/local/lib/python3.8/site-packages/crate/client/http.py&quot;, line 513, in _request
    self._drop_server(next_server, ex_message)
  File &quot;/usr/local/lib/python3.8/site-packages/crate/client/http.py&quot;, line 581, in _drop_server
    raise ConnectionError(
crate.client.exceptions.ConnectionError: No more Servers available, exception from last server: HTTPConnectionPool(host=&apos;crate-db&apos;, port=4200): Max retries exceeded with url: /_sql (Caused by NewConnectionError(&apos;&lt;urllib3.connection.HTTPConnection object at 0x7f29a923d340&gt;: Failed to establish a new connection: [Errno -2] Name or service not known&apos;))

Traceback (most recent call last):
  File &quot;./ls/sample_tasks.py&quot;, line 56, in &lt;module&gt;
    main()
  File &quot;./ls/sample_tasks.py&quot;, line 33, in main
    tasks_df = pd.read_sql(random_query, DATABASE_IP)
  File &quot;/usr/local/lib/python3.8/site-packages/pandas/io/sql.py&quot;, line 628, in read_sql
    return pandas_sql.read_query(
  File &quot;/usr/local/lib/python3.8/site-packages/pandas/io/sql.py&quot;, line 1579, in read_query
    result = self.execute(*args)
  File &quot;/usr/local/lib/python3.8/site-packages/pandas/io/sql.py&quot;, line 1424, in execute
    return self.connectable.execution_options().execute(*args, **kwargs)
  File &quot;&lt;string&gt;&quot;, line 2, in execute
  File &quot;/usr/local/lib/python3.8/site-packages/sqlalchemy/util/deprecations.py&quot;, line 401, in warned
    return fn(*args, **kwargs)
  File &quot;/usr/local/lib/python3.8/site-packages/sqlalchemy/engine/base.py&quot;, line 3139, in execute
    return connection.execute(statement, *multiparams, **params)
  File &quot;/usr/local/lib/python3.8/site-packages/sqlalchemy/engine/base.py&quot;, line 1274, in execute
    return self._exec_driver_sql(
  File &quot;/usr/local/lib/python3.8/site-packages/sqlalchemy/engine/base.py&quot;, line 1578, in _exec_driver_sql
    ret = self._execute_context(
  File &quot;/usr/local/lib/python3.8/site-packages/sqlalchemy/engine/base.py&quot;, line 1845, in _execute_context
    self._handle_dbapi_exception(
  File &quot;/usr/local/lib/python3.8/site-packages/sqlalchemy/engine/base.py&quot;, line 2026, in _handle_dbapi_exception
    util.raise_(
  File &quot;/usr/local/lib/python3.8/site-packages/sqlalchemy/util/compat.py&quot;, line 207, in raise_
    raise exception
  File &quot;/usr/local/lib/python3.8/site-packages/sqlalchemy/engine/base.py&quot;, line 1802, in _execute_context
    self.dialect.do_execute(
  File &quot;/usr/local/lib/python3.8/site-packages/sqlalchemy/engine/default.py&quot;, line 719, in do_execute
    cursor.execute(statement, parameters)
  File &quot;/usr/local/lib/python3.8/site-packages/crate/client/cursor.py&quot;, line 53, in execute
    self._result = self.connection.client.sql(sql, parameters,
  File &quot;/usr/local/lib/python3.8/site-packages/crate/client/http.py&quot;, line 396, in sql
    content = self._json_request(&apos;POST&apos;, self.path, data=data)
  File &quot;/usr/local/lib/python3.8/site-packages/crate/client/http.py&quot;, line 522, in _json_request
    response = self._request(method, path, data=data)
  File &quot;/usr/local/lib/python3.8/site-packages/crate/client/http.py&quot;, line 513, in _request
    self._drop_server(next_server, ex_message)
  File &quot;/usr/local/lib/python3.8/site-packages/crate/client/http.py&quot;, line 581, in _drop_server
    raise ConnectionError(
sqlalchemy.exc.OperationalError: (crate.client.exceptions.ConnectionError) No more Servers available, exception from last server: HTTPConnectionPool(host=&apos;crate-db&apos;, port=4200): Max retries exceeded with url: /_sql (Caused by NewConnectionError(&apos;&lt;urllib3.connection.HTTPConnection object at 0x7f29a923d340&gt;: Failed to establish a new connection: [Errno -2] Name or service not known&apos;))
[SQL:
        SELECT id, txt FROM example_data
        ORDER BY RANDOM()
        LIMIT 30;]
(Background on this error at: https://sqlalche.me/e/14/e3q8)
=== STARTING LABEL STUDIO TESTS WITH EXAMPLE DATA ===
Try sampling some tasks...
ERROR: Task sampling failed.
=== STARTING DATA PROGRAMMING TESTS WITH EXAMPLE DATA ===
Testing multiclass...
ERROR: Experiment run failed to complete.
=== STARTING MODELLING TESTS WITH EXAMPLE DATA ===
Testing multiclass with LM test data...
ERROR: Experiment run failed to complete.
b40e23f0ae4cba0c0f476f8490a04d4f4c5299cb72249cef17024f3debc46528
=== STARTING MODEL DEPLOYMENT TESTS WITH EXAMPLE DATA ===
Testing multiclass predictions...
{&quot;message&quot;:&quot;Oops! ExampleModel did not predict successfully. Please check that the requested data points have the \n            features corresponding to the features the model was trained on, and that the database is available.\n            Details: (crate.client.exceptions.ConnectionError) No more Servers available, exception from last server: HTTPConnectionPool(host=&apos;crate-db&apos;, port=4200): Max retries exceeded with url: /_sql (Caused by NewConnectionError(&apos;&lt;urllib3.connection.HTTPConnection object at 0x7fd46685e730&gt;: Failed to establish a new connection: [Errno -2] Name or service not known&apos;))\n[SQL: \n    SELECT id, txt_use FROM example_data WHERE id IN (&apos;18&apos;, &apos;24&apos;, &apos;22&apos;, &apos;19&apos;, &apos;15&apos;);\n    ]\n(Background on this error at: http://sqlalche.me/e/14/e3q8)&quot;}
ERROR: Getting predictions failed.
test-deploy-network
=== TESTING COMPLETE ===
</pre>

## Software

```
docker --version
Docker version 20.10.12-ce, build 459d0dfbbb51fb2423a43655e6c62368ec0f36c9

docker-compose --version
docker-compose version 1.29.2, build unknown

uname -r
5.13.8-191.current
```

## Hardware

<?xml version="1.0" encoding="utf-8"?>
<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN"
    "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
<html xmlns="http://www.w3.org/1999/xhtml">
<head>
<meta name="generator"  content="lshw-" />
<style type="text/css">
  .first {font-weight: bold; margin-left: none; padding-right: 1em;vertical-align: top; }
  .second {padding-left: 1em; width: 100%; vertical-align: center; }
  .id {font-family: monospace;}
  .indented {margin-left: 2em; border-left: dotted thin #dde; padding-bottom: 1em; }
  .node {border: solid thin #ffcc66; padding: 1em; background: #ffffcc; }
  .node-unclaimed {border: dotted thin #c3c3c3; padding: 1em; background: #fafafa; color: red; }
  .node-disabled {border: solid thin #f55; padding: 1em; background: #fee; color: gray; }
</style>
<title>aaron-x390</title>
</head>
<body>
<table width="100%" class="node" summary="attributes of aaron-x390">
 <thead><tr><td class="first">id:</td><td class="second"><div class="id">aaron-x390</div></td></tr></thead>
 <tbody>
    <tr><td class="first">description: </td><td class="second">Notebook</td></tr>
    <tr><td class="first">product: </td><td class="second">20Q1S02M00 (LENOVO_MT_20Q1_BU_Think_FM_ThinkPad X390)</td></tr>
    <tr><td class="first">vendor: </td><td class="second">LENOVO</td></tr>
    <tr><td class="first">version: </td><td class="second">ThinkPad X390</td></tr>
    <tr><td class="first">serial: </td><td class="second">PF1N577B</td></tr>
    <tr><td class="first">width: </td><td class="second">64 bits</td></tr>
    <tr><td class="first">capabilities: </td><td class="second"><dfn title="SMBIOS version 3.1.1">smbios-3.1.1</dfn> <dfn title="DMI version 3.1.1">dmi-3.1.1</dfn> <dfn title="Symmetric Multi-Processing">smp</dfn> <dfn title="32-bit processes">vsyscall32</dfn> </td></tr>
    <tr><td class="first">configuration:</td><td class="second"><table summary="configuration of aaron-x390"><tr><td class="sub-first"> administrator_password</td><td>=</td><td>disabled</td></tr><tr><td class="sub-first"> chassis</td><td>=</td><td>notebook</td></tr><tr><td class="sub-first"> family</td><td>=</td><td>ThinkPad X390</td></tr><tr><td class="sub-first"> power-on_password</td><td>=</td><td>disabled</td></tr><tr><td class="sub-first"> sku</td><td>=</td><td>LENOVO_MT_20Q1_BU_Think_FM_ThinkPad X390</td></tr><tr><td class="sub-first"> uuid</td><td>=</td><td>CCB1ACAE-E131-B211-A85C-FFF35FE54022</td></tr></table></td></tr>
 </tbody></table></div>
    <table width="100%" class="node" summary="attributes of core">
 <thead><tr><td class="first">id:</td><td class="second"><div class="id">core</div></td></tr></thead>
 <tbody>
       <tr><td class="first">description: </td><td class="second">Motherboard</td></tr>
       <tr><td class="first">product: </td><td class="second">20Q1S02M00</td></tr>
       <tr><td class="first">vendor: </td><td class="second">LENOVO</td></tr>
       <tr><td class="first">physical id: </td><td class="second"><div class="id">0</div></td></tr>
       <tr><td class="first">version: </td><td class="second">Not Defined</td></tr>
       <tr><td class="first">serial: </td><td class="second">L1HF92F00HB</td></tr>
       <tr><td class="first">slot: </td><td class="second">Not Available</td></tr>
 </tbody>    </table></div>
       <table width="100%" class="node" summary="attributes of memory">
 <thead><tr><td class="first">id:</td><td class="second"><div class="id">memory</div></td></tr></thead>
 <tbody>
          <tr><td class="first">description: </td><td class="second">System Memory</td></tr>
          <tr><td class="first">physical id: </td><td class="second"><div class="id">2</div></td></tr>
          <tr><td class="first">slot: </td><td class="second">System board or motherboard</td></tr>
          <tr><td class="first">size: </td><td class="second">16GiB</td></tr>
 </tbody>       </table></div>
          <table width="100%" class="node" summary="attributes of bank:0">
 <thead><tr><td class="first">id:</td><td class="second"><div class="id">bank:0</div></td></tr></thead>
 <tbody>
             <tr><td class="first">description: </td><td class="second">SODIMM DDR4 Synchronous 2667 MHz (0.4 ns)</td></tr>
             <tr><td class="first">product: </td><td class="second">M471A1K43BB1-CTD</td></tr>
             <tr><td class="first">vendor: </td><td class="second">Samsung</td></tr>
             <tr><td class="first">physical id: </td><td class="second"><div class="id">0</div></td></tr>
             <tr><td class="first">serial: </td><td class="second">00000000</td></tr>
             <tr><td class="first">slot: </td><td class="second">ChannelA-DIMM0</td></tr>
             <tr><td class="first">size: </td><td class="second">8GiB</td></tr>
             <tr><td class="first">width: </td><td class="second">64 bits</td></tr>
             <tr><td class="first">clock: </td><td class="second">2667MHz (0.4ns)</td></tr>
 </tbody>          </table></div>
          <table width="100%" class="node" summary="attributes of bank:1">
 <thead><tr><td class="first">id:</td><td class="second"><div class="id">bank:1</div></td></tr></thead>
 <tbody>
             <tr><td class="first">description: </td><td class="second">SODIMM DDR4 Synchronous 2667 MHz (0.4 ns)</td></tr>
             <tr><td class="first">product: </td><td class="second">M471A1K43BB1-CTD</td></tr>
             <tr><td class="first">vendor: </td><td class="second">Samsung</td></tr>
             <tr><td class="first">physical id: </td><td class="second"><div class="id">1</div></td></tr>
             <tr><td class="first">serial: </td><td class="second">00000000</td></tr>
             <tr><td class="first">slot: </td><td class="second">ChannelB-DIMM0</td></tr>
             <tr><td class="first">size: </td><td class="second">8GiB</td></tr>
             <tr><td class="first">width: </td><td class="second">64 bits</td></tr>
             <tr><td class="first">clock: </td><td class="second">2667MHz (0.4ns)</td></tr>
 </tbody>          </table></div>
       <table width="100%" class="node" summary="attributes of cache:0">
 <thead><tr><td class="first">id:</td><td class="second"><div class="id">cache:0</div></td></tr></thead>
 <tbody>
          <tr><td class="first">description: </td><td class="second">L1 cache</td></tr>
          <tr><td class="first">physical id: </td><td class="second"><div class="id">c</div></td></tr>
          <tr><td class="first">slot: </td><td class="second">L1 Cache</td></tr>
          <tr><td class="first">size: </td><td class="second">256KiB</td></tr>
          <tr><td class="first">capacity: </td><td class="second">256KiB</td></tr>
          <tr><td class="first">capabilities: </td><td class="second"><dfn title="Synchronous">synchronous</dfn> <dfn title="Internal">internal</dfn> <dfn title="Write-back">write-back</dfn> <dfn title="Unified cache">unified</dfn> </td></tr>
          <tr><td class="first">configuration:</td><td class="second"><table summary="configuration of cache:0"><tr><td class="sub-first"> level</td><td>=</td><td>1</td></tr></table></td></tr>
 </tbody>       </table></div>
       <table width="100%" class="node" summary="attributes of cache:1">
 <thead><tr><td class="first">id:</td><td class="second"><div class="id">cache:1</div></td></tr></thead>
 <tbody>
          <tr><td class="first">description: </td><td class="second">L2 cache</td></tr>
          <tr><td class="first">physical id: </td><td class="second"><div class="id">d</div></td></tr>
          <tr><td class="first">slot: </td><td class="second">L2 Cache</td></tr>
[…]
