### 2022-01-25 9412302
update manuscript and usage document
