+++ paper.bib
@@ -60,14 +60,18 @@ series = {DEEM'20}
 }
 
 @inproceedings{Gutierrez2021,
-abstract = {Social media platforms are public venues where conversations about issues of public interest take place. Much of the recent research has been devoted to evaluating the degree to which online conversations capture public opinion on issues of broad societal interest. We describe a robust and scalable platform to support such studies. Our platform allows the analysis of three semantic aspects of tweets, namely the personal values, sentiment, and humour expressed in them, as well as the public's engagement with them. In addition, it aggregates these indicators at the level of tweet authors to shed light on the activities and style of influencers of public opinion. Finally, it offers rich visualizations to enable users to gain insights on their datasets. We demonstrate the usefulness of our platform with two case studies: (a) analyzing the fragmented narratives around established (hydro, oil and gas, coal, nuclear) and new (solar, wind, geothermal, biomass) energy sources; and (b) comparing the social-media brands of academic institutions.},
-location = {Toronto, ON, Canada},
-author = {Gutierrez, Candelario Gutierrez and Whittaker, Andrea and Patenio, Katherine Mae and Gehman, Joel and Lefsrud, Lianne M and Barbosa, Denilson and Stroulia, Eleni},
-booktitle = {CASCON x EVOKE},
-keywords = {academic institutions' brands,energy narratives,humour,personal values,social media},
-publisher = {ACM Digital Library},
-title = {{Analyzing and Visualizing Twitter Conversations}},
-year = {2021}
+author = {Gutierrez, Candelario A. Gutierrez and Whittaker, Andrea and Patenio, Katherine Mae and Gehman, Joel and Lefsrud, Lianne M. and Barbosa, Denilson and Stroulia, Eleni},
+title = {Analyzing and Visualizing Twitter Conversations},
+year = {2021},
+publisher = {IBM Corp.},
+address = {USA},
+abstract = {Social media platforms are public venues where conversations about issues of public interest take place. Much recent research has been devoted to evaluating the degree to which online conversations capture public opinion on issues of broad societal interest. We describe a robust and scalable platform to support such studies. Our platform allows the analysis of three semantic aspects of tweets, namely the personal values, sentiment, and humor expressed in them, as well as the public's engagement with them. In addition, it aggregates these indicators at the level of tweet authors to shed light on the activities and style of influencers of public opinion. Finally, it offers rich visualizations to enable users to gain insights on their datasets. We demonstrate the usefulness of our platform with two case studies: (a) analyzing the fragmented narratives around established (hydro, oil and gas, coal, nuclear) and new (solar, wind, geothermal, biomass) energy sources; and (b) comparing the social-media brands of academic institutions.},
+booktitle = {Proceedings of the 31st Annual International Conference on Computer Science and Software Engineering},
+pages = {4–13},
+numpages = {10},
+keywords = {energy narratives, personal values, academic institutions' brands, humor, social media},
+location = {Toronto, Canada},
+series = {CASCON '21}
 }
 
 @misc{SnorkelFlow,
@@ -75,5 +79,23 @@ year = {2021}
 	url={https://snorkel.ai/platform/},
 	note={Proprietary software available from https://snorkel.ai/platform/},
 	author={Snorkel AI, Inc.},
-	year={2021},
+	year={2021}
+}
+
+@misc{Neuro,
+    title={{Neu.ro}},
+    url = {https://neu.ro/},
+    note = {MLOps Platform available from https://neu.ro/neuro-mlops/},
+    author = {Neu.ro},
+    year = {2022}
+}
+
+@misc{MLRun,
+  author = {Iguazio Ltd.},
[…]

[…]
