+++ paper.bib
@@ -91,11 +91,9 @@ series = {CASCON '21}
 }
 
 @misc{MLRun,
-  author = {Iguazio Ltd.},
+  author = {Iguazio, Ltd.},
   title = {{MLRun}},
   year = {2022},
-  publisher = {GitHub},
-  journal = {GitHub repository},
-  howpublished = {\url{https://github.com/mlrun/mlrun}},
+  url = {https://github.com/mlrun/mlrun},
   commit = {33ca54c34c40dc48dbe5d9108d690f20c2a83b87}
 }
