### 2022-01-25 U-Alberta/ADaPT-ML#9 @nulberry
Hey! I mentioned the issue about dependencies above because I believe that performing the bootstrap checks for CrateDB will address this issue. I made the mistake of not recognizing these bootstrap checks as a more essential part of system setup. In the code block you posted, it shows the `crate-db` container restarting, which is not supposed to happen:
```
Waiting for MLflow databases (5 sec)...
crate-db                  /docker-entrypoint.sh crat ...   Restarting
dp-mlflow-db              /entrypoint.sh mysqld            Up (health: starting)   3306/tcp, 33060/tcp, 33061/tcp
modelling-mlflow-db       /entrypoint.sh mysqld            Up (health: starting)   3306/tcp, 33060/tcp, 33061/tcp
Waiting for MLflow databases (5 sec)...
crate-db                  /docker-entrypoint.sh crat ...   Restarting
Waiting for MLflow databases (5 sec)...
Checking for exited containers...
Startup complete.
```

Please look into the bootstrap checks [here](https://crate.io/docs/crate/howtos/en/latest/admin/bootstrap-checks.html) and then try running the test script again.
