### 2022-01-26 U-Alberta/ADaPT-ML#9 @aaronpeikert
Running `docker run crate` resulted in a more informative failure:

```
[1] bootstrap checks failed
[1]: max virtual memory areas vm.max_map_count [65530] is too low, increase to at least [262144] by adding `vm.max_map_count = 262144` to `/etc/sysctl.conf` or invoking `sysctl -w vm.max_map_count=262144`
```
after `sysctl -w vm.max_map_count=262144`, `bash ./test/all-unix.sh` was successfull.
