### 2022-02-21 #4038 @wincowgerDEV
Thanks for the ping. I didn't realize that the software was only fully
fleshed out for Mac and Linux devices when I started because it wasn't
listed in the documentation materials. I don't have a Mac or Linux device
so I'm not sure if I should build a virtual machine to do the task or I
should request better documentation for Windows which the maintainer says
should be possible, though it is not fully fleshed out. We could also try
and add another reviewer who does have a Linux device and I can end my
review at this time with all the components done besides the beta test.

Warm Regards
Win

On Sun, Feb 20, 2022, 10:31 PM Jacob Schreiber ***@***.***>
wrote:

> Howdy @aaronpeikert <https://github.com/aaronpeikert> and @wincowgerDEV
> <https://github.com/wincowgerDEV>, how is the review going?
>
