### 2022-02-21 #4038 @wincowgerDEV
Aaron, that's good to know. It sounds to me like the documentation isn't
quite ready for prime time use yet and we should create an issue in the
repo to outline some of the things that are still left undone from a user
perspective because I don't think users would do the level of debugging
that we are doing right now to get it to work. There needs to be a
hold-your-hand level tutorial that gets everything up and running. I'll
start the issue shortly. I'll stay focused on the windows aspects for now
since there isn't a tutorial for using this with VMs. Though another thing
the maintainer could provide is an AMI or something to get this spooled up
immediately.

On Mon, Feb 21, 2022, 7:20 AM Aaron Peikert ***@***.***>
wrote:

> Hi, thanks for the reminder. I am a Linux user, but nevertheless had
> troubles with the setup that kinda dragged me down and made me forget that
> I still have to review. So I just needed a reminder to get back to it.
>
