### 2022-02-21 #4038 @aaronpeikert
Hi Win,
it is surely a good idea to improve Windows support. My issues were related to some dependency and not the project itself. My Linux setup is also not too common (only a few ten thousand users I think). I will test the project under some normal Linux and possibly under Mac OS and report back. My fear is that `docker-compose` is not as portable as advertised (I had the weirdest issues getting a container up and running somewhere else). A tricky problem for @U-Alberta to address. Having a  virtual machine solution available might work, but still can be hard for the target user group. Let's see how widespread the problem actually is, since documentation can only address common problems.
