### 2022-02-22 #4038 @wincowgerDEV
Thanks for the input Jacob. I'll see if the author has an interest in
broadening the support for Windows and if not will see if Aaron's able to
get it working on the options they have.

On Mon, Feb 21, 2022, 10:59 AM Jacob Schreiber ***@***.***>
wrote:

> Knowing that it's difficult to use the software is important. I agree that
> they should improve the documentation to make it clear what platforms are
> fully supported. I don't think it's necessary that they make it work on all
> platforms if that's particularly difficult. If one of you does not have
> access to the supported platforms I think it's okay for your review to be
> higher level, e.g., at the documentation and organizational level and on
> the paper itself. But it would be great if at least one of you could get
> the code working on a standard platform, at least in a minimal format.
>
