### 2022-02-24 b8de60a
add complete installation and testing instructions for Windows
