### 2022-02-24 #4038 @nulberry
Hello everyone,
Thank you for pointing out the issues with the documentation for Windows. I apologize for not having this documentation clear and accurate at the beginning of this review, and for any delay or frustration this may have caused. The latest commit has the improved documentation.

> Hi Win, it is surely a good idea to improve Windows support. My issues were related to some dependency and not the project itself. My Linux setup is also not too common (only a few ten thousand users I think). I will test the project under some normal Linux and possibly under Mac OS and report back. My fear is that `docker-compose` is not as portable as advertised (I had the weirdest issues getting a container up and running somewhere else). A tricky problem for @U-Alberta to address. Having a virtual machine solution available might work, but still can be hard for the target user group. Let's see how widespread the problem actually is, since documentation can only address common problems.

@aaronpeikert , I'm understanding from this comment that you are experiencing new issues that I am unaware of that are making it difficult for you to complete your review, is this correct? In https://github.com/U-Alberta/ADaPT-ML/issues/9 we addressed the issue caused by CrateDB failing the bootstrap checks, and you commented that the tests ran successfully. Can you please open a new issue detailing the problems you are having with `docker-compose`, and if you've determined how widespread the problem is?
