### 2022-02-25 #4038 @wincowgerDEV
@aaronpeikert I am running through the Windows set up so no need to do that
on your part.

On Fri, Feb 25, 2022 at 2:52 AM Aaron Peikert ***@***.***>
wrote:

> @aaronpeikert <https://github.com/aaronpeikert> , I'm understanding from
> this comment that you are experiencing new issues that I am unaware of that
> are making it difficult for you to complete your review, is this correct?
>
> No, on my personal machine, everything is fine. Nevertheless, I wanted to
> test your software on another Linux machine, on Mac, and on Windows (if you
> plan on supporting Windows). Just to verify that people can follow the
> documentation and do not run into any issues.
>
