### 2022-02-25 cab3c48
update installation and testing instructions, add test script for Windows
