+++ test/README.md
@@ -14,9 +14,7 @@ bash ./test/all-unix.sh
 
 ## Windows: Run all tests ##
 
-If you suspect that the `all-unix.sh` script will work on your host machine, then feel free to use it. Otherwise, the testing procedure can be broken down into these steps: 
-
-1. Open the Command Line in the ADaPT-ML root directory, and bring all the containers up:
+1. Open the Command Prompt in the ADaPT-ML root directory, and bring all the containers up:
 ```shell
 docker-compose --env-file .env --profile dev up -d
 ```
@@ -37,24 +35,9 @@ modelling-mlflow-db       "/entrypoint.sh mysq…"   m_db                running
 modelling-mlflow-deploy   "/start.sh"              m_deploy            running             0.0.0.0:8088->80/tcp
 modelling-mlflow-server   "mlflow server --bac…"   m_web               running             0.0.0.0:5001->5000/tcp
 ```
-3. Run the tests one-by-one:
+3. Run the test batch file:
 ```shell
-docker exec label-studio-dev python /test/ls-test.py
-```
-```shell
-docker exec dp-mlflow sh -c ". ~/.bashrc && python /test/dp-test.py"
-```
-```shell
-docker exec modelling-mlflow sh -c ". ~/.bashrc && python /test/ml-test.py"
-```
-```shell
-docker network create test-deploy-network --subnet 192.168.2.0/24 --gateway 192.168.2.10
-docker network connect --ip 192.168.2.4 test-deploy-network modelling-mlflow-deploy
-docker network connect --ip 192.168.2.8 test-deploy-network modelling-mlflow
-docker exec modelling-mlflow sh -c ". ~/.bashrc && python /test/deploy-test.py"
-docker network disconnect test-deploy-network modelling-mlflow-deploy
-docker network disconnect test-deploy-network modelling-mlflow
-docker network rm test-deploy-network
+test\all-windows.bat
 ```
 
 ## Optional Clean-up ##
@@ -64,3 +47,19 @@ If you wish to bring the containers down after testing, then:
 ```shell
 docker-compose down
 ```
+
+## Troubleshooting ##
+
+- If the `crate-db` container's status shows `Restarting`, please check its logs, `docker logs crate-db`. If you see something like this in the logs:
+    ```
+  [1] bootstrap checks failed
+  [1]: max virtual memory areas vm.max_map_count [65530] is too low, increase to at least [262144] by adding `vm.max_map_count = 262144` to `/etc/sysctl.conf` or invoking `sysctl -w vm.max_map_count=262144`
+    ```
+  then check that you completed the appropriate [bootstrap check steps](../usage.md#step-1-review-system-requirements).
+- If after the `docker-compose ... up -d` command, an error indicating that there is no space left occurs while pulling the images, then try these steps:
+  1. Check the amount of reclaimable space for images: `docker system df`
+  2. If reclaimable is greater than zero, clean up unused containers, networks, images, and the build cache: `docker system prune -a`
+  3. Try the testing procedure again from the beginning
+  4. If there still is not enough space, consider relocating the root directory ([Linux](https://linuxconfig.org/how-to-move-docker-s-default-var-lib-docker-to-another-directory-on-ubuntu-debian-linux), [Windows](https://dev.to/kimcuonthenet/move-docker-desktop-data-distro-out-of-system-drive-4cg2))
+
+For all other concerns with ADaPT-ML's tests, especially if any fail, please create an [Issue](https://github.com/U-Alberta/ADaPT-ML/issues).
