### 2022-03-03 a6ca9df
try installing Docker on MacOS runner. Update documentation
