+++ .github/workflows/ml-test.yml
@@ -9,10 +9,12 @@ jobs:
   build:
 
     runs-on: ${{ matrix.os }}
+
     strategy:
       fail-fast: false
       matrix:
-        os: [ ubuntu-latest, macos-latest, windows-latest ]
+        os: [ ubuntu-latest, macos-latest, windows-2022 ]
+
     defaults:
       run:
         shell: bash
@@ -26,6 +28,15 @@ jobs:
 
     steps:
     - uses: actions/checkout@v2
+
+    - name: If this is the MacOS runner, install Docker
+      if: ${{ runner.os == "macOS" }}
+      run: |
+        brew install docker docker-machine \
+        brew cask install virtualbox \
+        docker-machine create --driver virtualbox default \
+        docker-machine env default \
+        eval "$(docker-machine env default)"
 
     - name: Build the docker-compose profile and test modelling
       run: bash ./test/start-m.sh
