### 2022-03-03 #4038 @nulberry
To fully integrate Windows and MacOS into my CI, I am working on setting up `windows-2022` and `macos-latest` runners to perform the same testing that was running on `ubuntu-latest`. I have only basic experience with GitHub Workflows so it might take another day or two.
