+++ .github/workflows/dp-test.yml
@@ -23,7 +23,7 @@ jobs:
     - uses: actions/checkout@v2
 
     - name: If this is the MacOS runner, install Docker
-      if: matrix.os == "macos-latest"
+      if: matrix.os == 'macos-latest'
       run: |
         brew install docker docker-machine \
         brew cask install virtualbox \
