### 2022-03-03 12b4b41
try running tests on MacOS and Windows runners
