### 2022-03-03 35f8c9d
try running tests on MacOS and Windows runners
