+++ .github/workflows/dp-test.yml
@@ -24,26 +24,7 @@ jobs:
 
     - name: If this is the MacOS runner, install Docker
       if: matrix.os == 'macos-latest'
-      run: |
-        mkdir -p ~/.docker/machine/cache \
-        curl -Lo ~/.docker/machine/cache/boot2docker.iso https://github.com/boot2docker/boot2docker/releases/download/v19.03.12/boot2docker.iso \
-        brew install docker docker-machine \
-        docker-machine create --driver virtualbox default \
-        docker-machine env default \
-        eval "$(docker-machine env default)"
-
-#        brew install docker docker-machine \
-#        brew cask install virtualbox \
-#        docker-machine create --driver virtualbox default \
-#        docker-machine env default \
-#        eval "$(docker-machine env default)"
-
-#        HOMEBREW_NO_AUTO_UPDATE=1
-#        brew install --cask docker
-#        sudo /Applications/Docker.app/Contents/MacOS/Docker --unattended --install-privileged-components
-#        open -a /Applications/Docker.app --args --unattended --accept-license
-#        echo "We are waiting for Docker to be up and running. It can take over 2 minutes..."
-#        while ! /Applications/Docker.app/Contents/Resources/bin/docker info &>/dev/null; do sleep 1; done
+      run: bash ./test/install-docker.sh
 
     - name: Build the docker-compose profile and test data programming
       run: bash ./test/start-dp.sh
