+++ test/start-m.sh
@@ -1,6 +1,10 @@
 #!/bin/bash -e
 
-docker-compose --env-file .env --profile modelling up -d
+if [ "$RUNNER_OS" = "Windows" ]; then
+  docker-compose --file ./test/docker-compose-windows.yml --env-file .env --profile modelling up -d
+else
+  docker-compose --env-file .env --profile modelling up -d
+fi
 
 while docker-compose ps | grep -i "starting"
   do
