+++ test/start-deploy.sh
@@ -1,6 +1,11 @@
 #!/bin/bash -e
 
-docker-compose --env-file .env --profile deploy up -d
+if [ "$RUNNER_OS" = "Windows" ]; then
+  docker-compose --file ./test/docker-compose-windows.yml --env-file .env --profile deploy up -d
+else
+  docker-compose --env-file .env --profile deploy up -d
+fi
+
 while docker-compose ps | grep -i "starting"
   do
     echo "Waiting for MLflow databases (5 sec)..."
@@ -17,6 +22,14 @@ fi
 echo "Startup complete."
 docker-compose ps
 
+#if [ "$RUNNER_OS" == "Linux" ]; then
+#              apt install important_linux_software
+#         elif [ "$RUNNER_OS" == "Windows" ]; then
+#              choco install important_windows_software
+#         else
+#              echo "$RUNNER_OS not supported"
+#              exit 1
+#         fi
 # set up a temporary network to use to test the model deployment
 docker network create test-deploy-network --subnet 192.168.2.0/24 --gateway 192.168.2.10
 docker network connect --ip 192.168.2.4 test-deploy-network modelling-mlflow-deploy
