+++ test/docker-compose-windows.yml
@@ -1,0 +1,229 @@
+### THIS FILE ONLY TO BE USED FOR GITHUB CI ###
+
+version: '3.8'
+
+services:
+  cratedb:
+    image: crate:latest
+    container_name: crate-db
+    ports:
+      - "4200:4200"
+    networks:
+      - dp_network
+      - m_network
+      - ls_network
+      - deploy_network
+    volumes:
+      - ${DB_DATA_PATH}:/data
+      - ${DB_DATA_PATH}/repos:/repos
+    command: [ "crate", "-Cnetwork.host=_site_" , "-Cpath.repo=/repos"]
+    deploy:
+      replicas: 1
+      restart_policy:
+        condition: on-failure
+    environment:
+      - CRATE_HEAP_SIZE=2g
+
+  ls_web:
+    profiles:
+      - label
+      - dev
+    restart: on-failure
+    image: heartexlabs/label-studio:latest
+    container_name: label-studio-web
+    ports:
+      - "8080:8080"
+    networks:
+      - ls_network
+    volumes:
+      - ./label-studio/config:/config
+      - ${LS_DATA_PATH}:/label-studio/data
+      - ${LS_TASKS_PATH}:/tasks
+      - ${LS_ANNOTATIONS_PATH}:/annotations
+    environment:
+      - LABEL_STUDIO_LOCAL_FILES_SERVING_ENABLED=true
+      - LABEL_STUDIO_LOCAL_FILES_DOCUMENT_ROOT=/label-studio/files
+    command: [ "label-studio", "start" ]
+
+  ls:
+    profiles:
+      - label
+      - dev
+    restart: on-failure
+    image: nullberry/adapt-ml-ls:latest
+    container_name: label-studio-dev
+    networks:
+      - ls_network
+    volumes:
+      - ./label-studio/ls:/code/ls
+      - ${LS_DATA_PATH}:/label-studio/data
+      - ${LS_TASKS_PATH}:/tasks
+      - ${LS_ANNOTATIONS_PATH}:/annotations
+      - ./test:/test
+    environment:
+      - DATABASE_IP=${DATABASE_IP}
+    stdin_open: true
+    tty: true
+
+  dp_db:
+    profiles:
+      - data_programming
+      - dev
+    restart: on-failure
+    image: mysql/mysql-server:8.0.26
+    container_name: dp-mlflow-db
+    expose:
+      - 3306
+    networks:
+      - dp_network
+    volumes:
+      - ${DP_DATA_PATH}/dbdata:/var/lib/mysql
+    environment:
+      - MYSQL_DATABASE=${DP_MYSQL_DATABASE}
+      - MYSQL_USER=${DP_MYSQL_USER}
+      - MYSQL_PASSWORD=${DP_MYSQL_PASSWORD}
+      - MYSQL_ROOT_PASSWORD=${DP_MYSQL_ROOT_PASSWORD}
+
+  dp_web:
+    profiles:
+      - dev
+    restart: on-failure
+    image: nullberry/adapt-ml-mlflow-server:latest
+    container_name: dp-mlflow-server
+    ports:
+      - "5000:5000"
+    networks:
+      - dp_network
+    depends_on:
+      - dp_db
+    volumes:
+      - ${DP_DATA_PATH}/mlruns:/mlflow/mlruns
+    environment:
+      - MLFLOW_TRACKING_URI=${DP_MLFLOW_TRACKING_URI}
+    command: [ "mlflow",
+               "server",
+               "--backend-store-uri=$DP_MLFLOW_TRACKING_URI",
+               "--default-artifact-root=./mlruns",
+               "--host=0.0.0.0"]
+
+  dp:
[…]

[…]
