+++ test/start-m.sh
@@ -1,11 +1,11 @@
 #!/bin/bash -e
 
-if [ "$RUNNER_OS" = "Windows" ]; then
-  docker-compose --file ./test/docker-compose-windows.yml --env-file .env --profile modelling up -d
-else
-  docker-compose --env-file .env --profile modelling up -d
-fi
-
+#if [ "$RUNNER_OS" = "Windows" ]; then
+#  docker-compose --file ./test/docker-compose-windows.yml --env-file .env --profile modelling up -d
+#else
+#  docker-compose --env-file .env --profile modelling up -d
+#fi
+docker-compose --env-file .env --profile modelling up -d
 while docker-compose ps | grep -i "starting"
   do
     echo "Waiting for MLflow databases (5 sec)..."
