+++ test/docker-compose-windows.yml
@@ -4,6 +4,7 @@ version: '3.8'
 
 services:
   cratedb:
+    isolation: process
     image: crate:latest
     container_name: crate-db
     ports:
@@ -25,6 +26,7 @@ services:
       - CRATE_HEAP_SIZE=2g
 
   ls_web:
+    isolation: process
     profiles:
       - label
       - dev
@@ -46,6 +48,7 @@ services:
     command: [ "label-studio", "start" ]
 
   ls:
+    isolation: process
     profiles:
       - label
       - dev
@@ -66,6 +69,7 @@ services:
     tty: true
 
   dp_db:
+    isolation: process
     profiles:
       - data_programming
       - dev
@@ -85,6 +89,7 @@ services:
       - MYSQL_ROOT_PASSWORD=${DP_MYSQL_ROOT_PASSWORD}
 
   dp_web:
+    isolation: process
     profiles:
       - dev
     restart: on-failure
@@ -107,6 +112,7 @@ services:
                "--host=0.0.0.0"]
 
   dp:
+    isolation: process
     profiles:
       - data_programming
       - dev
@@ -132,6 +138,7 @@ services:
     tty: true
 
   m_db:
+    isolation: process
     profiles:
       - modelling
       - deploy
@@ -152,6 +159,7 @@ services:
       - MYSQL_ROOT_PASSWORD=${MODELLING_MYSQL_ROOT_PASSWORD}
 
   m_web:
+    isolation: process
     profiles:
       - dev
     restart: on-failure
@@ -174,6 +182,7 @@ services:
                "--host=0.0.0.0" ]
 
   m:
+    isolation: process
     profiles:
       - modelling
       - deploy
@@ -200,6 +209,7 @@ services:
     tty: true
 
   m_deploy:
+    isolation: process
     profiles:
       - deploy
       - dev
