### 2022-03-30 #4038 @jmschrei
@nulberry what branch is the main version, with revisions, of the paper on now? What is the version of the software, is there a tagged release, and is there a DOI (e.g. from Zenodo) that we can use?
