### 2022-03-30 #4038 @nulberry
:partying_face: The paper is on the main branch, and I just created a tagged release v1.0.0. I have not created a DOI for it.
