### 2022-03-31 #4038 @nulberry
It looks like I have to create a new release in order for Zenodo to create a DOI. Sorry I did this out of order, I've never done this before. Is it okay if I delete the release v1.0.0 and redo it?
