### 2022-03-31 #4038 @aaronpeikert
Alternatively, you can give the release a different tag (like `v1.0.0-tickle-zenodo`). But there is probably little reason for such precautions. If other things rely on your release on the other hand this would be the safe way (assuming the relying software knows to avoid non-semver releases, which most recognize). It is not unusual to see tags like `v1.4.8-docs` to retrigger the documentation ci.
