### 2022-03-31 #4038 @nulberry
This is good to know, thanks. I've deleted and redone the release, and now Zenodo has created a DOI: https://zenodo.org/badge/latestdoi/380004451
