### 2022-04-04 #4038 @kyleniemeyer
Hi @nulberry, could you edit the metadata of the Zenodo archive so that your name is spelled out (rather than using your GitHub username)? After that we should be ready to publish.
