### 2022-04-04 #4038 @nulberry
Hi! I created a `.zenodo.json` file in the repo to define the metadata and replaced `nulberry` with my full name. Is that all I have to do, or do I have to flip the switch on the Zenodo archive in order for it to apply my changes?
