### 2022-04-04 #4038 @kyleniemeyer
@nulberry you can actually just log into Zenodo and edit the metadata for that record
